Extract AuthorDeserializer from Storage class
