+++ src/main/java/storage/Storage.java
@@ -2,8 +2,10 @@ package storage;
 
 import author.Author;
 import author.AuthorList;
-
-import manga.Manga;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.gson.JsonParseException;
+import com.google.gson.reflect.TypeToken;
 
 import java.io.File;
 import java.io.FileReader;
@@ -11,16 +13,6 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.lang.reflect.Type;
 import java.util.logging.Logger;
-
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonArray;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonDeserializer;
-import com.google.gson.JsonDeserializationContext;
-import com.google.gson.JsonParseException;
-import com.google.gson.reflect.TypeToken;
 
 //@@author xenthm
 // singleton
@@ -111,23 +103,5 @@ public class Storage {
         } catch (IOException e) {
             logger.warning("Problems saving file, data will not be saved!" + e.getMessage());
         }
-    }
-}
-
-class AuthorDeserializer implements JsonDeserializer<Author> {
-    @Override
-    public Author deserialize(JsonElement json, Type typeOfAuthor, JsonDeserializationContext context)
-            throws JsonParseException {
-        JsonObject jsonObject = json.getAsJsonObject();
-        String authorName = jsonObject.get("authorName").getAsString();
-        Author author = new Author(authorName);
-
-        JsonArray mangaArray = jsonObject.getAsJsonArray("mangaList");
-        for (JsonElement mangaElement : mangaArray) {
-            Manga manga = context.deserialize(mangaElement, Manga.class);
-            manga.setAuthor(author);
-            author.addManga(manga);
-        }
-        return author;
     }
 }
